show previous suggestions when user is reviewing ideas.
Fixed a bug in NEXT_IDEA switch case in the activityReducer that was
causing the idea to be dropped from the next user's Q when the next user
is not the creator of the idea. The idea should be dropped when the next
user 'is' the creator of the idea.