Improve the logic to move ideas among users for review.
Earlier ideas were passed around in strict round robin fashion i.e. an
idea will move from user1 to user2 to user3 and so on. Even if the idea
is at the end of user2's Q and user4 has reviewed all his/her ideas,
he/she will have to wait until the idea makes it through user2 and
user3.

With the new logic, once a user reviews an idea, it is available for any
other user to pickup and review - no round robin.